Further discussion re: https://github.com/OctopusDeploy/Issues/issues/1643 resulted in a change of approach. A new variable was created (Octopus.Action.Azure.PackageExtractionPath). This is used by the Calamari Azure commands.
This reverts commit c2f89f90565ad4b101962ec09e21a2ca85466bc8, reversing
changes made to e0977877d402f1e2e45f6b888534824e8672b23b.
