+++ source/Calamari/Calamari.csproj
@@ -194,6 +194,7 @@
     <Compile Include="Deployment\Conventions\EnsureCloudServicePackageIsCtpFormatConvention.cs" />
     <Compile Include="Deployment\Conventions\ExtractAzureCloudServicePackageConvention.cs" />
     <Compile Include="Deployment\Conventions\ExtractPackageConvention.cs" />
+    <Compile Include="Deployment\Conventions\ExtractPackageToStagingDirectoryConvention.cs" />
     <Compile Include="Deployment\Conventions\FindCloudServicePackageConvention.cs" />
     <Compile Include="Deployment\Conventions\LogVariablesConvention.cs" />
     <Compile Include="Deployment\Conventions\FeatureScriptConvention.cs" />
@@ -203,6 +204,7 @@
     <Compile Include="Deployment\DeploymentStages.cs" />
     <Compile Include="Deployment\DeploymentWorkingDirectory.cs" />
     <Compile Include="Deployment\Conventions\PackagedScriptConvention.cs" />
+    <Compile Include="Deployment\Conventions\ExtractPackageToApplicationDirectoryConvention.cs" />
     <Compile Include="Deployment\Conventions\IConvention.cs" />
     <Compile Include="Deployment\Conventions\IInstallConvention.cs" />
     <Compile Include="Deployment\Conventions\IRollbackConvention.cs" />
