+++ source/Calamari/Deployment/Conventions/ExtractPackageToStagingDirectoryConvention.cs
@@ -1,0 +1,18 @@
+﻿using Calamari.Integration.FileSystem;
+using Calamari.Integration.Packages;
+
+namespace Calamari.Deployment.Conventions
+{
+    public class ExtractPackageToStagingDirectoryConvention : ExtractPackageConvention
+    {
+        public ExtractPackageToStagingDirectoryConvention(IPackageExtractor extractor, ICalamariFileSystem fileSystem) 
+            : base(extractor, fileSystem)
+        {
+        }
+
+        protected override string GetTargetPath(RunningDeployment deployment, PackageMetadata metadata)
+        {
+            return deployment.Variables[SpecialVariables.Action.Azure.PackageExtractionPath];
+        }
+    }
+}
