+++ source/Calamari/Deployment/SpecialVariables.cs
@@ -120,6 +120,8 @@ namespace Calamari.Deployment
 
             public static class Azure
             {
+                public static readonly string PackageExtractionPath = "Octopus.Action.Azure.PackageExtractionPath";
+
                 public static readonly string SubscriptionId = "Octopus.Action.Azure.SubscriptionId";
                 public static readonly string CertificateBytes = "Octopus.Action.Azure.CertificateBytes";
                 public static readonly string CertificateThumbprint = "Octopus.Action.Azure.CertificateThumbprint";
